third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLineElement-dom-x1-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x1' attribute of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.getAttribute('x1') is "100"
PASS lineElement.getAttribute('x1') is "10"
PASS successfullyParsed is true

TEST COMPLETE
